requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.3
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
whitenoise==6.6.0